完成會員端 寵物各項功能 新增寵物 修改寵物(類彈窗方式) 刪除寵物(軟刪除)

+++ backend/src/main/java/com/petlife/model/Pet.java
@@ -1,6 +1,6 @@
 package com.petlife.model;
 
-import java.math.BigDecimal;
+
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -34,7 +34,7 @@ public class Pet {
 	    private Integer age;
 
 	    @Column(name = "weight")
-	    private BigDecimal weight;
+	    private Double weight;
 
 	    @Column(name = "medical_history")
 	    private String medicalHistory;
